requirements.txt
django
djangorestframework
environs
requests
django-easy-validator
pyjwt
django-cors-headers